tests: Move exif information bdd step to conftest
Like this it is accessible to all tests in the image directory,
especially also to the metadata test which can make good use of it.
